Extracted id function to lib file
